+++ Ass5/src/main/java/com/example/demojpa/repository/ProductRepository.java
@@ -11,9 +11,13 @@ import org.springframework.data.jpa.repository.Query;
 @Repository
 public interface ProductRepository extends JpaRepository<Product, Long> {
 
-    // Modify this query to support pagination and sorting.
-    @Query("SELECT p FROM Product p WHERE LOWER(p.name) LIKE LOWER(CONCAT('%', :keyword, '%')) AND p.price <= :maxPrice")
-    Page<Product> searchProducts(@Param("keyword") String keyword, @Param("maxPrice") double maxPrice, Pageable pageable);
+    // JPQL query hỗ trợ tìm kiếm theo keyword (name/description), categoryId, minPrice, maxPrice, có phân trang
+    @Query("SELECT p FROM Product p WHERE (LOWER(p.name) LIKE LOWER(CONCAT('%', :keyword, '%')) OR LOWER(p.description) LIKE LOWER(CONCAT('%', :keyword, '%'))) AND (:categoryId IS NULL OR p.category.id = :categoryId) AND p.price BETWEEN :minPrice AND :maxPrice")
+    Page<Product> searchProducts(@Param("keyword") String keyword,
+                                 @Param("categoryId") Long categoryId,
+                                 @Param("minPrice") double minPrice,
+                                 @Param("maxPrice") double maxPrice,
+                                 Pageable pageable);
 
     @Query(value = "SELECT COUNT(*) FROM product WHERE category_id = :categoryId", nativeQuery = true)
     long countProductsByCategoryId(@Param("categoryId") Long categoryId);
