+++ Ass5/src/main/java/com/example/demojpa/service/CustomerService.java
@@ -1,0 +1,12 @@
+package com.example.demojpa.service;
+
+public interface CustomerService {
+
+    // Define methods for customer-related operations here
+    // For example:
+    // Page<CustomerDTO> getAllCustomers(Pageable pageable);
+    // Optional<CustomerDTO> getCustomerById(Long id);
+    // CustomerDTO createCustomer(CustomerDTO dto);
+    // Optional<CustomerDTO> updateCustomer(Long id, CustomerDTO dto);
+    // void deleteCustomer(Long id);
+}
