+++ Ass5/src/main/java/com/example/demojpa/service/UserService.java
@@ -9,43 +9,13 @@ import java.util.List;
 import java.util.Optional;
 
 @Service
-public class UserService {
-    @Autowired
-    private UserRepository userRepository;
+public interface UserService {
 
-    public List<User> getAllUsers() {
-        return userRepository.findAll();
-    }
-
-    public Optional<User> getUserById(Long id) {
-        return userRepository.findById(id);
-    }
-
-    public User createUser(User user) {
-        return userRepository.save(user);
-    }
-
-    public void deleteUser(Long id) {
-        if (!userRepository.findById(id).isPresent()) {
-            throw new RuntimeException("User not found");
-        }
-        userRepository.deleteById(id);
-    }
-
-    public User updateUser(Long id, User userDetails) {
-        return userRepository.findById(id).map(user -> {
-            user.setUsername(userDetails.getUsername());
-            user.setEmail(userDetails.getEmail());
-            user.setPassword(userDetails.getPassword());
-            return userRepository.save(user);
-        }).orElseThrow(() -> new RuntimeException("User not found"));
-    }
-
-    public User findByUsername(String username) {
-        return userRepository.findByUsername(username);
-    }
-
-    public User findByEmail(String email) {
-        return userRepository.findByEmail(email);
-    }
+    List<User> getAllUsers();
+    Optional<User> getUserById(Long id);
+    User createUser(User user);
+    void deleteUser(Long id);
+    User updateUser(Long id, User userDetails);
+    User findByUsername(String username);
+    User findByEmail(String email);
 }
