+++ Ass5/src/main/java/com/example/demojpa/service/impl/CustomerServiceImpl.java
@@ -1,0 +1,5 @@
+package com.example.demojpa.service.impl;
+
+public interface CustomerServiceImpl {
+
+}
